requirements_fabrice.txt
certifi
charset-normalizer
contourpy
cycler
Cython
docopt
elastic-transport
elasticsearch
et-xmlfile
fonttools
happybase
hdfs
idna
kiwisolver
matplotlib
numpy
openpyxl
packaging
pandas
pillow
ply
pyparsing
python-dateutil
pytz
requests
setuptools
six
thriftpy
tzdata
Unidecode
urllib3
watchdog
zipp
inference_sdk
